Create README.md
Contributing to theHarvester Project

Welcome to theHarvester project, so you would like to contribute. The following below must be met to get accepted.
CI

Make sure all CI passes and you do not introduce any alerts from lgtm.
Unit Tests

For new modules a unit test for that module is required and we use pytest.
Coding Standards

    No single letter variables and variable names must represent the action that it is performing
    Have static typing on functions etc
    Make sure no errors are reported from mypy
    No issues reported with flake8

Submitting Bugs

If you find a bug in a module that you want to submit an issue for and know how to write python code. Please create a unit test for that bug(If possible) and submit a fix for it as it would be a big help to the project.